Moved versions tests, reactivated tests for protocol version
