+++ tests/protocols/gmpv214/system/test_versions.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-# Copyright (C) 2019-2021 Greenbone Networks GmbH
+# Copyright (C) 2021 Greenbone Networks GmbH
 #
 # SPDX-License-Identifier: GPL-3.0-or-later
 #
@@ -16,19 +16,16 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-from . import Gmpv208TestCase
-from .testcmds import *  # pylint: disable=unused-wildcard-import, wildcard-import
+from ...gmpv214 import Gmpv214TestCase
+from ...gmpv208.system.versions import GmpGetVersionTestCase
+from .versions import GmpGetProtocolVersionTestCase
 
 
-class Gmpv208HelpTestCase(GmpHelpTestCase, Gmpv208TestCase):
+class Gmpv214GetVersionCommandTestCase(GmpGetVersionTestCase, Gmpv214TestCase):
     pass
 
 
-class Gmpv208GetSystemReportsTestCase(
-    GmpGetSystemReportsTestCase, Gmpv208TestCase
+class Gmpv214GmpGetProtocolVersionTestCase(
+    GmpGetProtocolVersionTestCase, Gmpv214TestCase
 ):
-    pass
-
-
-class Gmpv208v7WithStatementTestCase(GmpWithStatementTestCase, Gmpv208TestCase):
     pass
